+++ config/bootstrap.php
@@ -1,28 +1,17 @@
 <?php
 /**
- * This file is part of MeCms.
+ * This file is part of me-cms.
  *
- * MeCms is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as
- * published by the Free Software Foundation, either version 3 of the
- * License, or (at your option) any later version.
+ * Licensed under The MIT License
+ * For full copyright and license information, please see the LICENSE.txt
+ * Redistributions of files must retain the above copyright notice.
  *
- * MeCms is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with MeCms.  If not, see <http://www.gnu.org/licenses/>.
- *
- * @author      Mirko Pagliai <mirko.pagliai@gmail.com>
- * @copyright   Copyright (c) 2016, Mirko Pagliai for Nova Atlantis Ltd
- * @license     http://www.gnu.org/licenses/agpl.txt AGPL License
- * @link        http://git.novatlantis.it Nova Atlantis Ltd
+ * @copyright   Copyright (c) Mirko Pagliai
+ * @link        https://github.com/mirko-pagliai/me-cms
+ * @license     https://opensource.org/licenses/mit-license.php MIT License
  */
-/**
- * (here `Cake\Core\Plugin` is used, as the plugins are not yet all loaded)
- */
+
+// (here `Cake\Core\Plugin` is used, as the plugins are not yet all loaded)
 use Cake\Core\Configure;
 use Cake\Core\Plugin;
 use Cake\Network\Request;
@@ -30,21 +19,15 @@ use Cake\Routing\DispatcherFactory;
 
 $request = new Request;
 
-/**
- * Requires the base of bootstrap
- */
+//Requires the base of bootstrap
 require_once __DIR__ . DS . 'bootstrap_base.php';
 
-/**
- * Loads DebugKit, if debugging is enabled
- */
+//Loads DebugKit, if debugging is enabled
 if (getConfig('debug') && !Plugin::loaded('DebugKit')) {
     Plugin::load('DebugKit', ['bootstrap' => true]);
 }
 
-/**
- * Loads other plugins
- */
+//Loads other plugins
 Plugin::load('Thumber', ['bootstrap' => true, 'routes' => true]);
 Plugin::load('Tokens', ['bootstrap' => true]);
 Plugin::load('DatabaseBackup', ['bootstrap' => true]);