+++ src/Model/Behavior/TreeBehavior.php
@@ -1,29 +1,20 @@
 <?php
 /**
- * This file is part of MeCms.
+ * This file is part of me-cms.
  *
- * MeCms is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as
- * published by the Free Software Foundation, either version 3 of the
- * License, or (at your option) any later version.
+ * Licensed under The MIT License
+ * For full copyright and license information, please see the LICENSE.txt
+ * Redistributions of files must retain the above copyright notice.
  *
- * MeCms is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with MeCms.  If not, see <http://www.gnu.org/licenses/>.
- *
- * @author      Mirko Pagliai <mirko.pagliai@gmail.com>
- * @copyright   Copyright (c) 2016, Mirko Pagliai for Nova Atlantis Ltd
- * @license     http://www.gnu.org/licenses/agpl.txt AGPL License
- * @link        http://git.novatlantis.it Nova Atlantis Ltd
+ * @copyright   Copyright (c) Mirko Pagliai
+ * @link        https://github.com/mirko-pagliai/me-cms
+ * @license     https://opensource.org/licenses/mit-license.php MIT License
  * @see         http://api.cakephp.org/3.4/class-Cake.ORM.Behavior.TreeBehavior.html
  */
 namespace MeCms\Model\Behavior;
 
 use Cake\ORM\Behavior\TreeBehavior as CakeTreeBehavior;
+use Cake\ORM\Query;
 
 /**
  * Makes the table to which this is attached to behave like a nested set and
@@ -31,11 +22,6 @@ use Cake\ORM\Behavior\TreeBehavior as CakeTreeBehavior;
  *  derived hierarchical structure.
  *
  * This behavior rewrites the `TreeBehavior` class provided by CakePHP.
- *
- * To add to your table:
- * <code>
- * $this->addBehavior('MeCms.Tree');
- * </code>
  */
 class TreeBehavior extends CakeTreeBehavior
 {
@@ -50,7 +36,7 @@ class TreeBehavior extends CakeTreeBehavior
      * @see http://api.cakephp.org/3.4/class-Cake.ORM.Behavior.TreeBehavior.html#_findTreeList
      * @uses Cake\ORM\Behavior\TreeBehavior::findTreeList()
      */
-    public function findTreeList(\Cake\ORM\Query $query, array $options)
+    public function findTreeList(Query $query, array $options)
     {
         $options['spacer'] = empty($options['spacer']) ? '—' : $options['spacer'];