+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # 2.x branch
 ## 2.19 branch
 ### 2.19.3
-* added initial schema of the plugin database;
+* added initial schema of the plugin database. Removed `BaseUpdateConsole` and
+    `UpdateShell` classes;
 * added `getConfigOrFail()` global function;
 * `firstImage()` is no longer a global function, but a method provided by the
     `GetPreviewFromTextTrait` class.