+++ CHANGELOG.md
@@ -5,7 +5,8 @@
     `UpdateShell` classes;
 * added `getConfigOrFail()` global function;
 * `firstImage()` is no longer a global function, but a method provided by the
-    `GetPreviewFromTextTrait` class.
+    `GetPreviewFromTextTrait` class;
+* the MIT license has been applied.
 
 ### 2.19.2
 * fixed bug in the list of active pages for each page category;